requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
atomicwrites==1.4.0
attrs==20.3.0
chromedriver-binary-auto==0.1.1
colorama==0.4.4
glob2==0.7
iniconfig==1.1.1
install==1.3.4
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
parse==1.19.0
parse-type==0.5.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-bdd==4.0.2
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.4
